move query screen request to session utils
