+++ src/test/java/org/commcare/backend/suite/model/test/QueryModelTests.java
@@ -2,6 +2,7 @@ package org.commcare.backend.suite.model.test;
 
 import com.google.common.collect.ImmutableMap;
 
+import org.cli.MockSessionUtils;
 import org.commcare.core.encryption.CryptUtil;
 import org.commcare.core.interfaces.MemoryVirtualDataInstanceStorage;
 import org.commcare.data.xml.VirtualInstances;
@@ -10,10 +11,8 @@ import org.commcare.suite.model.RemoteQueryDatum;
 import org.commcare.suite.model.SessionDatum;
 import org.commcare.test.utilities.CaseTestUtils;
 import org.commcare.test.utilities.MockApp;
-import org.commcare.util.mocks.MockQueryClient;
 import org.commcare.util.screen.CommCareSessionException;
 import org.commcare.util.screen.QueryScreen;
-import org.javarosa.core.model.instance.ExternalDataInstance;
 import org.jetbrains.annotations.NotNull;
 import org.junit.Assert;
 import org.junit.Before;
@@ -84,16 +83,12 @@ public class QueryModelTests {
         Assert.assertEquals("registry1", datum.getDataId());
 
         // construct the screen
-
-        QueryScreen screen = new QueryScreen(
-                "username", "password",
-                System.out, virtualDataInstanceStorage);
-        screen.init(session);
-
-
         // mock the query response
         InputStream response = this.getClass().getResourceAsStream("/case_claim_example/query_response.xml");
-        screen.setClient(new MockQueryClient(response));
+        QueryScreen screen = new QueryScreen(
+                "username", "password",
+                System.out, virtualDataInstanceStorage, new MockSessionUtils(response));
+        screen.init(session);
         return screen;
     }
 }
